Show file index and total file on the title bar.  Adjust gallery behavior.
Allow file selection in Image Manager.  Main Window controller behavior now is if it is selected with a single file, it will load the file directory and show the selected file.  If it is loaded with a set of files, the set will be used as a gallery and show the first file.